// reality/cloud/xrhome/src/client/common/types/db.ts
type UserAccountRole = "OWNER" | "ADMIN" | "DEV" | "BILLMANAGER"